Optimized main loop, fixed in-world menu bug
I have applied to large optimizations to the main loop to make it run more efficiently. Firstly, the program will now check if one of the menu states has changed overall and act upon that, rather than checking each. Secondly I have organized tickone and ticktwo, and I have changed the priorities of the the igm and the time scale (I swapped them).
Lastly, I have resolved the bug where the combat menus (and alike) are uncontrollable by limiting the frametime to the minMenus variable.